requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
boltons @ file:///Users/ec2-user/ci_py311/boltons_1678396074811/work
brotlipy==0.7.0
cachetools==5.3.3
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b64zphacdv/croot/certifi_1683875375103/work/certifi
cffi @ file:///Users/ec2-user/ci_py311/cffi_1678315312775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
conda @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_5aeztnykre/croot/conda_1689269911902/work
conda-content-trust @ file:///Users/ec2-user/ci_py311/conda-content-trust_1678409283937/work
conda-libmamba-solver @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b7jucd8kvj/croot/conda-libmamba-solver_1685032330942/work/src
conda-package-handling @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_f61sssj1s3/croot/conda-package-handling_1685024797154/work
conda_package_streaming @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_78ue32m8ay/croot/conda-package-streaming_1685019689310/work
contourpy==1.2.1
cryptography @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d3qwb8u102/croot/cryptography_1686613065937/work
cycler==0.12.1
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna @ file:///Users/ec2-user/ci_py311/idna_1678315819133/work
iniconfig==2.0.0
Jinja2==3.1.3
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
libmambapy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_91xegtksvi/croot/mamba-split_1685993177539/work/libmambapy
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
numpy==1.26.4
packaging @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c3zlhpboab/croot/packaging_1678965318579/work
pandas==2.1.1
pillow==10.3.0
pluggy @ file:///Users/ec2-user/ci_py311/pluggy_1678315400971/work
protobuf==4.25.3
pyarrow==15.0.2
pycosat @ file:///Users/ec2-user/ci_py311/pycosat_1678378899646/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.1b0
Pygments==2.17.2
pyOpenSSL @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_19gfn1ib_u/croot/pyopenssl_1678965300171/work
pyparsing==3.1.2
PySocks @ file:///Users/ec2-user/ci_py311/pysocks_1678315868424/work
pytest==7.4.2
python-dateutil==2.8.2
pytz==2024.1
referencing==0.34.0
requests @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ec2sxfa_zk/croot/requests_1682607519671/work
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml @ file:///Users/ec2-user/ci_py311/ruamel.yaml_1678379291841/work
seaborn==0.13.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz @ file:///Users/ec2-user/ci_py311/toolz_1678323009669/work
tornado==6.4
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_03acd6t6ca/croot/tqdm_1679561866522/work
typing_extensions==4.11.0
tzdata==2024.1
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1968zekhce/croot/urllib3_1686163172153/work
zstandard @ file:///Users/ec2-user/ci_py311_2/zstandard_1679338593428/work
